requirements.txt
amqp==5.3.1
asgiref==3.8.1
attrs==25.3.0
autopep8==2.3.2
billiard==4.2.1
build==1.2.2.post1
CacheControl==0.14.2
celery==5.5.1
certifi==2025.1.31
charset-normalizer==3.4.1
cleo==2.1.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
crashtest==0.4.1
distlib==0.3.9
Django==5.1.7
django-axes==7.0.2
django-cors-headers==4.3.1
django-eventstream==5.3.2
django-filter==25.1
django-grip==3.5.2
django-silk==5.3.2
django-storages==1.14.6
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
dotenv==0.9.9
drf-spectacular==0.27.1
dulwich==0.22.7
envparse==0.2.0
fastjsonschema==2.21.1
filelock==3.17.0
gprof2dot==2025.4.14
gripcontrol==4.2.0
idna==3.10
inflection==0.5.1
installer==0.7.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keyring==25.6.0
kombu==5.5.2
MarkupSafe==3.0.2
more-itertools==10.6.0
msgpack==1.1.0
mypy==1.15.0
mypy-extensions==1.0.0
packaging==24.2
pkginfo==1.12.1.2
platformdirs==4.3.6
poetry==2.0.1
poetry-core==2.0.1
prompt_toolkit==3.0.50
pubcontrol==3.5.0
pycodestyle==2.13.0
PyJWT==2.10.1
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
pywin32-ctypes==0.2.3
PyYAML==6.0.2
RapidFuzz==3.12.1
referencing==0.36.2
requests==2.32.3
requests-toolbelt==1.0.0
rpds-py==0.24.0
shellingham==1.5.4
six==1.17.0
sqlparse==0.5.3
tomlkit==0.13.2
trove-classifiers==2025.2.18.16
typing_extensions==4.12.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.3.0
vine==5.1.0
virtualenv==20.29.2
wcwidth==0.2.13
Werkzeug==3.1.3
whitenoise==6.11.0
